feat-chore: openai_audio package to manage code
