+++ autobots/action/action_type/action_audio2text/action_audio2text_transcription_openai.py
@@ -6,7 +6,7 @@ from pydantic import HttpUrl, ValidationError, BaseModel
 from autobots.action.action_type.abc.IAction import IAction, ActionConfigType, ActionInputType, ActionOutputType
 from autobots.action.action_type.action_types import ActionType
 from autobots.conn.openai.openai_client import get_openai
-from autobots.conn.openai.transcription_model import TranscriptionReq
+from autobots.conn.openai.openai_audio.transcription_model import TranscriptionReq
 from autobots.core.log import log
 
 
@@ -39,7 +39,7 @@ class ActionAudio2TextTranscriptionOpenai(IAction[TranscriptionReq, AudioRes, Tr
             if action_input.url is not None:
                 self.action_config.file_url = HttpUrl(action_input.url)
 
-            transcription = await get_openai().transcription(self.action_config)
+            transcription = await get_openai().openai_audio.transcription(self.action_config)
             return transcription
         except ValidationError as e:
             log.exception(e)
