+++ autobots/api/v1/api_actions/audio2text/api_actions_audio2text_transcription_openai.py
@@ -11,7 +11,7 @@ from autobots.action.action.user_actions import UserActions
 from autobots.action.action_type.action_audio2text.action_audio2text_transcription_openai import AudioRes
 from autobots.action.action_type.action_types import ActionType
 from autobots.auth.security import get_user_from_access_token
-from autobots.conn.openai.transcription_model import TranscriptionReq
+from autobots.conn.openai.openai_audio.transcription_model import TranscriptionReq
 from autobots.core.database.mongo_base import get_mongo_db
 from autobots.core.log import log
 from autobots.user.user_orm_model import UserORM
