+++ autobots/conn/openai/openai_audio/speech_model.py
