+++ autobots/conn/openai/openai_audio/transcription_model.py
