+++ autobots/conn/openai/openai_audio/translation_model.py
