+++ tests/action/test_action_audio2text_translation_openai.py
@@ -8,8 +8,8 @@ from autobots.action.action_type.action_audio2text.action_audio2text_translation
     ActionAudio2TextTranslationOpenai
 from autobots.conn.aws.aws_s3 import get_public_s3
 from autobots.conn.openai.openai_client import get_openai
-from autobots.conn.openai.speech_model import SpeechReq
-from autobots.conn.openai.translation_model import TranslationReq
+from autobots.conn.openai.openai_audio.speech_model import SpeechReq
+from autobots.conn.openai.openai_audio.translation_model import TranslationReq
 from autobots.core.utils import gen_hash
 
 
@@ -19,7 +19,7 @@ async def test_action_audio2text_translation_openai_happy_path(set_test_settings
     openai_client = get_openai()
     input = f"Namaste, main AutobotX hoon. main aapaki sahaayata ke lie yahaan hoon. Aaj ki tareekh {datetime.datetime.now().strftime('%B %d, %Y')} hai."
     params = SpeechReq(input=input)
-    resp:  HttpxBinaryResponseContent = await openai_client.speech(speech_req=params)
+    resp:  HttpxBinaryResponseContent = await openai_client.openai_audio.speech(speech_req=params)
     assert resp is not None
     url = await get_public_s3().put_file_obj(io.BytesIO(resp.content),f"{gen_hash(input)}.{params.response_format}")
     assert url != ""
